Species search logs correct return data
